+++ bitbybit/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## bitbybit 1.3.0 (breaking change)
+
+### Added
+
+### Changed
+
+- bitenum's new_with_raw_value, when used with an arbitrary-int type used to return the next larger whole primitive integer, which is unexpected. Slight change, but possibility for breakage:
+```rs
+// Old type
+let foo: Result<MyEnum, u8> = MyEnum::new_with_raw_value(u2::new(0b10));
+
+// New type (notice the u2 instead of u8)
+let foo: Result<MyEnum, u2> = MyEnum::new_with_raw_value(u2::new(0b10));
+```
+
+
+### Fixed
+
 ## bitbybit 1.2.2
 
 ### Added